Create docker data volume for snapweb
To be able to distribute the web client in more flavors, lets create a
data volume that can be shared.

To be able to use the container, a persistent volume is used to be
shared with the snapserver container.

Simply starting the container with a volume argument will populate the
volume. Due to the nature of how persistent volume works, updating the
volume, requires to delete/rename the original one and create a new one.

It is up to the user how to so. See the documentation for some examples.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>